# api/swagger.yaml
basePath: /
definitions:
  db.OrderRecipe:
    properties:
      created_at:
        type: string
      employee_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      recipe_id:
        type: string
      removed_at:
        type: string
    type: object
...
  httpx.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.APIError'
    type: object
...
  orders.AttachRecipeRequest:
    properties:
      employeeId:
        type: string
      recipeId:
        type: string
    type: object
  orders.ComponentView:
    properties:
      brand:
        type: string
      flavour:
        type: string
      percent:
        type: integer
    type: object
  orders.FeedbackView:
    properties:
      orderRecipeId:
        type: string
      review:
        type: string
      score:
        type: integer
    type: object
  orders.OpenRequest:
    properties:
      restaurantId:
        type: string
      tableId:
        example: "12"
        type: string
      userId:
        type: string
    type: object
  orders.OrderRecipeView:
    properties:
      authorFullName:
        type: string
      authorShortName:
        type: string
      components:
        items:
          $ref: '#/definitions/orders.ComponentView'
        type: array
      isSecret:
        type: boolean
      orderRecipeId:
        type: string
      recipeId:
        type: string
      recipeName:
        type: string
      strength:
        type: integer
    type: object
  orders.OrderResponse:
    properties:
      closedAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      recipes:
        items:
          $ref: '#/definitions/orders.OrderRecipeView'
        type: array
      restaurantId:
        type: string
      tableId:
        type: string
      userId:
        type: string
    type: object
  orders.RateOrderRecipeRequest:
    properties:
      score:
        example: 5
        type: integer
      userId:
        type: string
    type: object
  orders.ReviewOrderRecipeRequest:
    properties:
      review:
        type: string
      userId:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend for the mixMaster restaurant ordering / recipes app.
  title: mixMaster API
  version: 0.1.0
paths:
...
  /order-recipes/{orderRecipeId}/rating:
    post:
      consumes:
      - application/json
      parameters:
      - description: OrderRecipe ID
        in: path
        name: orderRecipeId
        required: true
        type: string
      - description: User + score
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orders.RateOrderRecipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/orders.FeedbackView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Rate an order recipe (1..5)
      tags:
      - orders
  /order-recipes/{orderRecipeId}/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: OrderRecipe ID
        in: path
        name: orderRecipeId
        required: true
        type: string
      - description: User + review
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orders.ReviewOrderRecipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/orders.FeedbackView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Leave a text review on an order recipe
      tags:
      - orders
  /orders/{id}/close:
    post:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/orders.OrderResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Reset table (delete TableAssignment + set closed_at)
      tags:
      - orders
  /orders/{id}/recipes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Recipe + author
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orders.AttachRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/db.OrderRecipe'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "409":
          description: order is closed
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Attach a recipe to an order
      tags:
      - orders
  /orders/{id}/recipes/{orderRecipeId}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: OrderRecipe ID
        in: path
        name: orderRecipeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Remove/replace a recipe on an order (soft, removed_at)
      tags:
      - orders
  /orders/open:
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant + table
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orders.OpenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: existing order
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/orders.OrderResponse'
              type: object
        "201":
          description: new order
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/orders.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Open a table order (get-or-create, idempotent)
      tags:
      - orders
...
swagger: "2.0"