# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
)

type Component struct {
	ID        uuid.UUID `json:"id"`
	RecipeID  uuid.UUID `json:"recipe_id"`
	Brand     string    `json:"brand"`
	Flavour   string    `json:"flavour"`
	Percent   int32     `json:"percent"`
	CreatedAt time.Time `json:"created_at"`
}

type Employee struct {
	ID         uuid.UUID `json:"id"`
	FirstName  string    `json:"first_name"`
	LastName   string    `json:"last_name"`
	MiddleName string    `json:"middle_name"`
	ShortName  string    `json:"short_name"`
	CreatedAt  time.Time `json:"created_at"`
	UpdatedAt  time.Time `json:"updated_at"`
}

type EmployeeRating struct {
	ID         uuid.UUID `json:"id"`
	EmployeeID uuid.UUID `json:"employee_id"`
	UserID     uuid.UUID `json:"user_id"`
	Score      int32     `json:"score"`
	CreatedAt  time.Time `json:"created_at"`
	UpdatedAt  time.Time `json:"updated_at"`
}

type EmployeeRestaurant struct {
	EmployeeID   uuid.UUID `json:"employee_id"`
	RestaurantID uuid.UUID `json:"restaurant_id"`
	Rating       *int32    `json:"rating"`
	CreatedAt    time.Time `json:"created_at"`
	Position     *string   `json:"position"`
}

type Favourite struct {
	UserID        uuid.UUID `json:"user_id"`
	OrderRecipeID uuid.UUID `json:"order_recipe_id"`
	LikedAt       time.Time `json:"liked_at"`
}

type MenuRecipe struct {
	ID               uuid.UUID  `json:"id"`
	RestaurantID     uuid.UUID  `json:"restaurant_id"`
	AuthorEmployeeID uuid.UUID  `json:"author_employee_id"`
	Name             string     `json:"name"`
	Description      string     `json:"description"`
	Strength         int32      `json:"strength"`
	Price            float64    `json:"price"`
	Rating           *float64   `json:"rating"`
	Tags             []string   `json:"tags"`
	RemovedAt        *time.Time `json:"removed_at"`
	CreatedAt        time.Time  `json:"created_at"`
	UpdatedAt        time.Time  `json:"updated_at"`
	Badge            *string    `json:"badge"`
}

type Order struct {
	ID           uuid.UUID  `json:"id"`
	TableID      string     `json:"table_id"`
	RestaurantID uuid.UUID  `json:"restaurant_id"`
	UserID       *uuid.UUID `json:"user_id"`
	CreatedAt    time.Time  `json:"created_at"`
	ClosedAt     *time.Time `json:"closed_at"`
}

type OrderRecipe struct {
	ID         uuid.UUID  `json:"id"`
	OrderID    uuid.UUID  `json:"order_id"`
	RecipeID   uuid.UUID  `json:"recipe_id"`
	EmployeeID uuid.UUID  `json:"employee_id"`
	CreatedAt  time.Time  `json:"created_at"`
	RemovedAt  *time.Time `json:"removed_at"`
}

type OrderRecipeFeedback struct {
	ID            uuid.UUID `json:"id"`
	OrderRecipeID uuid.UUID `json:"order_recipe_id"`
	UserID        uuid.UUID `json:"user_id"`
	Score         *int32    `json:"score"`
	Review        *string   `json:"review"`
	CreatedAt     time.Time `json:"created_at"`
	UpdatedAt     time.Time `json:"updated_at"`
}

type Recipe struct {
	ID        uuid.UUID `json:"id"`
	Name      *string   `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	Strength  *int32    `json:"strength"`
	IsSecret  bool      `json:"is_secret"`
}

type Restaurant struct {
	ID        uuid.UUID `json:"id"`
	Name      string    `json:"name"`
	Code      string    `json:"code"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Shift struct {
	ID           uuid.UUID `json:"id"`
	RestaurantID uuid.UUID `json:"restaurant_id"`
	EmployeeID   uuid.UUID `json:"employee_id"`
	ShiftDate    time.Time `json:"shift_date"`
	CreatedAt    time.Time `json:"created_at"`
}

type TableAssignment struct {
	ID           uuid.UUID `json:"id"`
	RestaurantID uuid.UUID `json:"restaurant_id"`
	TableID      string    `json:"table_id"`
	OrderID      uuid.UUID `json:"order_id"`
	UpdatedAt    time.Time `json:"updated_at"`
}

type User struct {
	ID          uuid.UUID `json:"id"`
	PhoneNumber string    `json:"phone_number"`
	Gender      *string   `json:"gender"`
	CreatedAt   time.Time `json:"created_at"`
}